+++ store/store.ts
@@ -39,25 +39,6 @@ export interface CreateStore<S extends AnyState> {
   middleware?: BaseMiddleware[];
 }
 
-export function createBatchMdw<S extends AnyState>(
-  queue: (send: () => void) => void = queueMicrotask,
-) {
-  let notifying = false;
-  return function* batchMdw(_: UpdaterCtx<S>, next: Next) {
-    try {
-      if (!notifying) {
-        notifying = true;
-        yield* action<void>(function* (resolve) {
-          queue(resolve);
-        });
-        yield* next();
-      }
-    } finally {
-      notifying = false;
-    }
-  };
-}
-
 export function createStore<S extends AnyState>({
   initialState,
   scope: initScope,
